# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "AppController_getHealthStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "Healthy"
                  }
                }
              }
            }
          }
        },
        "summary": "Checks if the service is healthy",
        "tags": [
          "Health"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User signup payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User login payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/user": {
      "get": {
        "operationId": "UserController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    }
  },
  "info": {
    "title": "Snap Talk",
    "description": "This is a simple API for demonstration purposes",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "greet",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "SignupRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the user",
            "example": "johndoe"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "mypassword"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the user",
            "example": "johndoe"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "mypassword"
          }
        },
        "required": [
          "username",
          "password"
        ]
      }
    }
  }
}